Release halogen 6

+++ Halogen/src/texel.h
@@ -7,5 +7,5 @@ void Texel(std::vector<int*> params);
 void LoadBench4Per(std::vector<std::pair<Position, double>>& positions, std::string& line, int lineCount, Position& position);
 void Loadquietlabeled(std::vector<std::pair<Position, double>>& positions, std::string& line, int lineCount, Position& position);
 void PrintIteration(double error, std::vector<int*>& params, std::vector<double> paramiterValues, double step_size, int iteration);
-double CalculateError(std::vector<std::pair<Position, double>>& positions, SearchData& data, double k, unsigned int subset);
+double CalculateError(std::vector<std::pair<Position, double>>& positions, SearchData& data, double k, size_t subset);
 double CalculateK(std::vector<Position>& positionList, std::vector<double>& positionScore, std::vector<double>& positionResults);
